Move build url to sugarlib
