+++ sugarlib/helpers.py
@@ -2,6 +2,7 @@ import humanize
 
 from datetime import datetime
 from sugarlib.redis_helpers import r_master_etag
+from urllib import parse
 
 
 def humanize_delta(date_value):
@@ -35,3 +36,21 @@ def get_expires_on_ttl(expires_datetime):
     if ttl < 0:
         return False
     return int(ttl)
+
+
+def build_url(url: str, relative_url: str = "", query_params: dict = {}) -> str:
+    """Build absolute url
+    Args:
+            url: URL path
+            relative_url: Relative URL path
+            query_params (dict): Query params for the url. Defaults to {}
+    """
+    parsed_url = parse.urlparse(url)
+    parsed_query = parse.parse_qs(parsed_url.query)
+    query_params.update(parsed_query)
+    url = parsed_url.scheme + "://" + parsed_url.netloc
+
+    absoulte_url = parse.urljoin(url, relative_url)
+    if query_params:
+        absoulte_url = absoulte_url + "?" + parse.urlencode(query_params, doseq=True)
+    return absoulte_url
